@@ -3,30 +3,30 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
     "@reduxjs/toolkit": "^1.9.1",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
-    "axios": "^0.21.1",
-    "cloudinary-react": "^1.6.8",
-    "lodash": "^4.17.20",
+    "axios": "^1.2.2",
+    "cloudinary-react": "^1.8.1",
+    "lodash": "^4.17.21",
     "middleware": "^1.0.0",
     "moment": "^2.29.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-google-login": "^5.2.2",
-    "react-hook-form": "^7.10.1",
-    "react-icons": "^4.1.0",
+    "react-hook-form": "^7.42.1",
+    "react-icons": "^4.7.1",
     "react-loader-spinner": "^5.3.4",
-    "react-redux": "^7.2.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
-    "reactjs-popup": "^2.0.4",
-    "redux": "^4.0.5",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.6.2",
+    "react-scripts": "^5.0.1",
+    "reactjs-popup": "^2.0.5",
+    "redux": "^4.2.0",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.3.0"
+    "redux-thunk": "^2.4.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -47,7 +47,7 @@
       "last 1 chrome version",
       "last 1 firefox version",
       "last 1 safari version"
-    ],
-    "proxy": "http://localhost: 3000/*"
-  }
+    ]
+  },
+  "proxy": "http://localhost: 3000/*"
 }
