@@ -1,9 +1,9 @@
-import { Route, Switch } from "react-router-dom";
+import { Route, Routes } from "react-router-dom";
 import SignUpParent from "./components/SignUpParent";
 import SignUpChild from "./components/SignUpChild";
 import AppEntry from "./components/AppEntry";
 import Login from "./components/Login";
-import childLogin from "./components/ChildLogin";
+import ChildLogin from "./components/ChildLogin";
 import AddChore from "./components/AddChore";
 import UpdateChore from "./components/UpdateChore";
 import ChildList from "./components/ChildList";
@@ -16,29 +16,25 @@ import Home from "./components/Home";
 import UpdateChild from "./components/UpdateChild";
 
 const App = () => {
+  const token = localStorage.getItem('token');
   return (
-    <Switch>
-      <Route exact path="/" component={AppEntry} />
-      <Route path="/login" component={Login} />
-      <Route path="/signUpParent" component={SignUpParent} />
-      <Route path="/childLogin" component={childLogin} />
-      <PrivateRoute path="/home" component={Home} />
-      <PrivateRoute path="/signUpChild" component={SignUpChild} />
-      <PrivateRoute path="/updateChild/:id" component={UpdateChild} />
-      <PrivateRoute path="/addchore" component={AddChore} />
-      <PrivateRoute path="/updateChore" component={UpdateChore} />
-      <PrivateRoute path="/childList" component={ChildList} />
-      <PrivateRoute path="/chore/:id" component={Chore} />
-      <PrivateRoute path="/choreList/:id" component={ChoreList} />
-      <PrivateRoute path="/child/:id" component={Child} />
-      <PrivateRoute path="/childdetail/:id">
-        {" "}
-        <ChildDetail />
-      </PrivateRoute>
-      <PrivateRoute path="/updatechild/:id">
-        <UpdateChild />
-      </PrivateRoute>
-    </Switch>
+    <Routes>
+      <Route exact path="/" element={<AppEntry/>} />
+      <Route path="/login" element={<Login/>} />
+      <Route path="/signUpParent" element={<SignUpParent/>} />
+      <Route path="/childLogin" element={<ChildLogin/>} />
+      <Route element={<PrivateRoute token={token} />} >
+        <Route path="/home" element={<Home/>}/>
+        <Route path="/childList" element={<ChildList/>}/>
+        <Route path="/signUpChild" element={<SignUpChild/>} />        <Route path="/child/:id" element={<Child/>} />
+        <Route path="/childdetail/:id" element={<ChildDetail />}/>
+        <Route path="/updatechild/:id" element={<UpdateChild/>}/>
+        <Route path="/chore/:id" element={<Chore/>} />
+        <Route path="/choreList/:id" element={<ChoreList/>} />
+        <Route path="/addchore" element={<AddChore/>} />
+        <Route path="/updateChore" element={<UpdateChore/>} />       
+      </Route>
+    </Routes>
   );
 };
 
