@@ -2,7 +2,7 @@ import Avatar from "@material-ui/core/Avatar";
 import Button from "@material-ui/core/Button";
 import CssBaseline from "@material-ui/core/CssBaseline";
 import TextField from "@material-ui/core/TextField";
-import { Link, useHistory } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import Box from "@material-ui/core/Box";
 import Typography from "@material-ui/core/Typography";
 import CardContent from "@material-ui/core/CardContent";
@@ -83,7 +83,7 @@ const ChildForm = ({
   value,
   changeHandler,
 }) => {
-  const history = useHistory();
+  const history = useNavigate();
 
   const classes = useStyles();
 
