@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import Avatar from "@material-ui/core/Avatar";
 import Button from "@material-ui/core/Button";
 import CssBaseline from "@material-ui/core/CssBaseline";
@@ -40,7 +40,7 @@ const useStyles = makeStyles((theme) => ({
 
 export default function SignUpChild () {
 
-  const history = useHistory();
+  const history = useNavigate();
   const dispatch = useDispatch()
 
   const [state, setState] = useState({
