@@ -1,17 +1,14 @@
 import React from 'react';
-import {Route, Redirect} from 'react-router-dom';
+import { Navigate, Outlet } from 'react-router-dom';
 
-export const PrivateRoute = ({component: Component, ...rest}) => {
-  return (
-    <Route
-      {...rest}
-      render={props => {
-        if (localStorage.getItem('token')) {
-          return <Component {...props} />;
-        } else {
-          return <Redirect to='/login' />;
-        }
-      }}
-    />
-  );
+export const PrivateRoute = ({ 
+  token,
+  redirectPath = '/',
+  children
+ }) => {
+
+  if(!token) {
+    return <Navigate to={redirectPath} replace />
+  }
+  return children ? children : <Outlet />
 };
