@@ -1,19 +1,22 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import * as ReactDOM from 'react-dom/client'
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
 import store from './store'
 import './index.css';
 import App from './App';
 
+const container = document.getElementById("root");
+const root = ReactDOM.createRoot(container);
 
-ReactDOM.render(
+
+root.render(
 <React.StrictMode>
   <Provider store={store}>
     <Router>
       <App />
     </Router>
   </Provider>
-</React.StrictMode>,
-  document.getElementById('root')
+</React.StrictMode>
+
 );
